App-Impact-Reporting/requirements.txt
numpy==1.26.0
pandas==2.0.3
python-dateutil==2.8.2
python-dotenv==0.21.0
sqlalchemy==2.0.20
psycopg2-binary==2.9.3
flask==2.3.3
openpyxl==3.0.10
gunicorn==21.2.0
Flask-CORS==3.0.10